final release

+++ src/meanvariant.rs
@@ -9,18 +9,18 @@ use std::error::Error;
  Date: 2025-5-15
 */
 
-pub fn meanquality(hashvector: Vec<Genomecapture>) -> Result<usize, Box<dyn Error>> {
-    let mut meanquality: Vec<usize> = Vec::new();
+pub fn meanquality(hashvector: Vec<Genomecapture>) -> Result<f64, Box<dyn Error>> {
+    let mut meanquality: Vec<f64> = Vec::new();
     let hashvector_borrow = hashvector;
     for i in hashvector_borrow.iter() {
-        meanquality.push(i.qual.clone().parse::<usize>().unwrap());
+        meanquality.push(i.qual.clone().parse::<f64>().unwrap());
     }
-    let mut meanvariant: usize = 0usize;
+    let mut meanvariant: f64 = 0.0 as f64;
     for i in meanquality.iter() {
         meanvariant += i;
     }
     let finallen = meanquality.len();
-    Ok(meanvariant / finallen as usize)
+    Ok(meanvariant / finallen as f64)
 }
 
 /*
